# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PatrolX API
  version: 3.0.0
  description: PatrolX 巡检系统接口契约（离线巡检：任务/系统/规则/字典）。 当前契约包含 /api/v2 巡检接口和 /api/v3 KPI 指标库接口。
servers:
- url: /
paths:
...
  /api/v2/tasks:
    post:
      summary: 创建巡检任务（同名同 checksum 复用；同名不同 checksum 冲突）
      operationId: createTaskV2
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - package_file
              properties:
                package_file:
                  type: string
                  format: binary
                  description: 数据压缩包（zip/tar.gz，非空，默认上限 2GB）
                name:
                  type: string
                  description: 任务名称（可选）
                version:
                  type: string
                  description: 版本字典编码（可选）
                province:
                  type: string
                  description: 省份字典编码（可选）
                operator:
                  type: string
                  description: 运营商字典编码（可选）
                product:
                  type: string
                  description: 产品形态字典编码（可选）
      responses:
        '202':
          description: 任务已创建，异步执行
          headers:
            Location:
              schema:
                type: string
                example: /tasks/task-001
              example: /api/v2/tasks/task-001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreatedV2'
        '400':
          $ref: '#/components/responses/Error400'
        '409':
          description: 同名包但 checksum 不同
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorV2'
        '413':
          $ref: '#/components/responses/Error413'
    get:
      summary: 任务列表
      operationId: listTasksV2
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/TaskStatusV2'
      responses:
        '200':
          description: 任务摘要列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponseV2'
        '422':
          $ref: '#/components/responses/Error422'
  /api/v2/tasks/{task_id}:
    get:
      summary: 任务摘要（不含全量规则结果）
      operationId: getTaskV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 任务摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionTaskV2'
        '404':
          $ref: '#/components/responses/Error404'
    delete:
      summary: 删除任务（级联删除 uploads/ 与 output/ 现场数据）
      operationId: deleteTaskV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: 已删除
        '409':
          description: 任务正在排队或执行，禁止删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorV2'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          description: 任务现场删除失败；任务记录与可恢复现场保持可见
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDeleteErrorV2'
  /api/v2/tasks/{task_id}/rerun:
    post:
      summary: 增量重跑（全部或指定规则）
      operationId: rerunTaskV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RerunRequestV2'
      responses:
        '202':
          description: 重跑任务已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreatedV2'
        '404':
          $ref: '#/components/responses/Error404'
  /api/v2/tasks/{task_id}/report:
    get:
      summary: 巡检报告（HTML 在线预览）
      operationId: getReportV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: HTML 报告
          content:
            text/html:
              schema:
                type: string
          headers: {}
        '404':
          $ref: '#/components/responses/Error404'
  /api/v2/tasks/{task_id}/logs:
    get:
      summary: 任务执行日志
      operationId: getTaskLogsV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 结构化执行日志
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskLogsV2'
        '404':
          $ref: '#/components/responses/Error404'
  /api/v2/tasks/{task_id}/system:
    get:
      summary: 系统摘要与规则状态列表
      operationId: getSystemV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      - name: exclude_details
        in: query
        schema:
          type: boolean
          default: false
        description: 为 true 时只返回规则状态摘要，清空 rules[].metadata/metrics/findings。
      responses:
        '200':
          description: 系统巡检结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInspectionV2'
        '404':
          $ref: '#/components/responses/Error404'
  /api/v2/tasks/{task_id}/rules/{rule_code}:
    get:
      summary: 单规则完整结果
      operationId: getRuleResultV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      - $ref: '#/components/parameters/RuleCode'
      - name: exclude_records
        in: query
        schema:
          type: boolean
          default: false
        description: 为 true 时清空 KPI metadata.kpi_files[].records，并对 kpi_results[].series 与 provenance.direct_cross_reference 做最多 200 点的展示抽稀；完整序列保留在规则结果文件中。
      responses:
        '200':
          description: 规则结果（metrics/findings）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResultV2'
        '404':
          $ref: '#/components/responses/Error404'
  /api/v2/tasks/{task_id}/rules/{rule_code}/kpi/records:
    get:
      summary: 分页查询 KPI 原始记录
      operationId: listKpiRecordsV2
      parameters:
      - $ref: '#/components/parameters/TaskId'
      - $ref: '#/components/parameters/RuleCode'
      - name: metric_key
        in: query
        schema:
          type: string
        description: 稳定指标 key，精确匹配。
      - name: source_file
        in: query
        schema:
          type: string
        description: 任务内相对路径，精确匹配。
      - name: period_minutes
        in: query
        schema:
          type: integer
          enum: [5, 15, 30, 60]
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/KpiDisplayStatusV2'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/KpiRecordPageSize'
      responses:
        '200':
          description: KPI 原始记录分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KpiRecordPageV2'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
...
  /api/v2/inspectors:
    get:
      summary: 已注册规则元数据（只读）
      operationId: listInspectorsV2
      parameters:
      - name: category
        in: query
        schema:
          $ref: '#/components/schemas/RuleCategoryV2'
      - name: include_hidden
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: 规则元数据列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InspectorInfoV2'
...
  /api/v2/dicts/{dict_name}:
    put:
      summary: 维护字典项（纯内部开放）
      operationId: updateDictV2
      parameters:
      - name: dict_name
        in: path
        required: true
        schema:
          type: string
          enum:
          - province
          - operator
          - product
          - version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DictUpdateRequestV2'
      responses:
        '200':
          description: 更新后的字典
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DictItemV2'
        '400':
          $ref: '#/components/responses/Error400'
...
  /api/v3/tasks/{task_id}/kpi/catalog-snapshot:
    get:
      summary: 读取任务 KPI 配置快照
      operationId: getKpiCatalogSnapshotV3
      parameters:
      - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 任务 KPI 配置快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KpiTaskCatalogSnapshotV3'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          description: 任务存在但 KPI 配置快照缺失
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorV2'
        '500':
          $ref: '#/components/responses/Error500'
components:
  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
    RuleCode:
      name: rule_code
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    KpiRecordPageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
  responses:
    Error400:
      description: 请求错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error413:
      description: 文件过大
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorV2'
    Error500:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorV2'
  schemas:
...
    TaskStatusV2:
      type: string
      enum:
      - pending
      - running
      - completed
      - failed
...
    RuleCategoryV2:
      type: string
      enum:
      - log
      - kpi
      - traffic
      - alarm
      - config
      - resource
      - other
...
    RuleResultV2:
      type: object
      required:
      - code
      - name
      - category
      - priority
      - execution_order
      - status
      - severity
      properties:
        code:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/RuleCategoryV2'
        priority:
          $ref: '#/components/schemas/PriorityV2'
        execution_order:
          type: integer
        status:
          $ref: '#/components/schemas/RuleStatusV2'
        severity:
          $ref: '#/components/schemas/SeverityV2'
        summary:
          type: string
        skip_reason:
          type: string
          description: status=skip 时必填
        executed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/MetricV2'
        findings:
          type: array
          items:
            $ref: '#/components/schemas/FindingV2'
        metadata:
          type: object
          additionalProperties: true
    SystemInspectionV2:
      type: object
      required:
      - package_file
      - status
      - summary
      - rules
      properties:
        package_file:
          type: string
        package_checksum:
          type: string
        version:
          type: string
        status:
          type: string
          enum:
          - completed
          - failed
        summary:
          $ref: '#/components/schemas/SummaryV2'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/RuleResultV2'
        customer:
          type: object
          additionalProperties: true
...
    InspectionTaskV2:
      type: object
      required:
      - task_id
      - name
      - mode
      - status
      - trigger
      - created_at
      - stats
      properties:
        task_id:
          type: string
        name:
          type: string
        mode:
          $ref: '#/components/schemas/TaskModeV2'
        status:
          $ref: '#/components/schemas/TaskStatusV2'
        trigger:
          $ref: '#/components/schemas/TaskTriggerV2'
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        stats:
          $ref: '#/components/schemas/TaskStatsV2'
        preparation:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/DataPreparationV2'
        system:
          $ref: '#/components/schemas/SystemInspectionV2'
    TaskCreatedV2:
      type: object
      required:
      - task_id
      properties:
        task_id:
          type: string
...
    TaskListResponseV2:
      type: object
      required:
      - items
      - total
      - page
      - page_size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskSummaryV2'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    RerunRequestV2:
      type: object
      properties:
        rule_codes:
          type: array
          description: 不填表示重跑该任务全部规则
          items:
            type: string
...
    TaskLogsV2:
      type: object
      required:
      - task_id
      - entries
      properties:
        task_id:
          type: string
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LogEntryV2'
...
    DictItemV2:
      type: object
      required:
      - code
      - name
      properties:
        code:
          type: string
          description: 编码，仅小写字母/数字/下划线
        name:
          type: string
...
    DictUpdateRequestV2:
      type: object
      required:
      - code
      - name
      properties:
        code:
          type: string
        name:
          type: string
    TaskDeleteErrorV2:
      allOf:
      - $ref: '#/components/schemas/ErrorV2'
      - type: object
        properties:
          detail:
            $ref: '#/components/schemas/TaskDeleteErrorDetailV2'
      description: 任务删除失败错误（detail 携带重试上下文）
...
    ErrorV2:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - invalid_package
          - package_too_large
          - invalid_dict
          - bad_request
          - unknown_rule
          - not_found
          - internal
          - task_delete_failed
          - package_checksum_conflict
          - invalid_filename
          - invalid_resource_csv
          - resource_conflict
          - corrupt_data
          - validation_error
          description: 错误码（新增时同步契约与实现）
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
      description: 错误码（/api/v2 含包上传与损坏数据错误）
    KpiDisplayStatusV2:
      type: string
      enum:
      - pass
      - warn
      - fail
      - neutral
      - unavailable
...
    KpiRecordPageV2:
      type: object
      required: [total, page, page_size, items]
      properties:
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 200
        items:
          type: array
          items:
            $ref: '#/components/schemas/KpiRecordItemV2'
...
    KpiTaskCatalogSnapshotV3:
      type: object
      required:
      - schema_version
      - base_data_version
      - classification_version
      - captured_at
      - metrics
      - rules
      properties:
        schema_version:
          type: integer
          enum:
          - 1
        base_data_version:
          type: string
        classification_version:
          type: integer
          minimum: 0
        captured_at:
          type: string
          format: date-time
        metrics:
          type: array
          items:
            type: object
            additionalProperties: true
        rules:
          type: object
          additionalProperties: true